Update checklist.md
I added this line because this is the second time a partner has stopped the entire pull request process due to thinking they need to test every single command. Hopefully, future partners will read that line and know that is not the case. @ShahafBenYakir let me know if this is fine or if there is a better way to word it. Thank you!!! :)