Article for the phishing campaign (#700)
* Create Phishing-Campaign.md

* Update Phishing-Campaign.md

Edited and reviewed

* Update Phishing-Campaign.md

Added architecture diagram, better explained how similarity is determined

* Update Phishing-Campaign.md

Added a link to the Phishing Tutorial (in the first paragraph).

* Update Phishing-Campaign.md

Added canvas

* Update Phishing-Campaign.md

Updated the content with new screenshots

* Update Phishing-Campaign.md

Removed lower similarity part. irrelevant for this release

* Update Phishing-Campaign.md

Updated article with changes that were and fixed wrong playbook name

* Update Phishing-Campaign.md

* Update Phishing-Campaign.md

Updated the doc so that it is based on the template. Can you check that I have all of the incident fields? The playbook link won't work until it is published and then I will update it.

* Update Phishing-Campaign.md

Updated with Ido's changes

Co-authored-by: Richard Bluestone <53567272+richardbluestone@users.noreply.github.com>