XSOAR CI/CD Article (#882)
* CI/CD Article

* Add ID Title and Description

* Some fixes

* build fixes

* build fixes

* Add files via upload

* Update CI-CD-Article.md

* Update CI-CD-Article.md

Fixed spacing in step 7.5

* Fixed broken images

* Fixed broken images

* Fixed broken images

* Fixed broken images

* Fixed broken images

* Update CI-CD-Article.md

Restructured and added additional content. It might be better to use google docs as its easier to track. Let me know if you want me to add it also to google docs.

* Update CI-CD-Article.md

Moving the MT note about the CI/CD Content Pack to the Artifact section in deployment.

* Some fixes

* Some fixes

* CRs and some fixes

* fix md

* Update CI-CD-Article.md

Removing the content in the brackets - Upload the content via the CICD build.

* Update CI-CD-Article.md

Removed the second option in the Note for MT

* change image

Co-authored-by: Richard Bluestone <53567272+richardbluestone@users.noreply.github.com>