Add note on `CommandResults` `readable_output` arg
Small change: added a note in the row of the `CommandResults` table for the `readable_output` arg to specify how to create a War Room entry only, and not write any data to context, because I had a client ask about this. Feel free to rephrase or move the note to a different part of the page, as you see fit.