Update content-management.md
Change this
"You can deploy your content without the need for an artifact server. If you do not use an artifact server, the demisto-sdk upload command, uploads the content pack directly to Marketplace."

To this:
"You can deploy your content without the need for an artifact server. If you do not use an artifact server, the `demisto-sdk upload` command uploads the content pack directly to your XSOAR instance."